tools/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
snowflake-snowpark-python

